moved demo to own file
